+++ examples/windows/triangle.cc
@@ -19,10 +19,13 @@
 class Triangle : public vs::Application {
   void OnStart() override {
     // clang-format off
-    float vertices[] = {
-      -0.5f, -0.5f, 0.0f,
-       0.5f, -0.5f, 0.0f,
-       0.0f,  0.5f, 0.0f,
+    vertices_ = {
+      {{-0.5f, -0.5f, 0.0f}, {{ 0.0f,  0.0f,  0.0f}}},
+      {{ 0.5f, -0.5f, 0.0f}, {{ 0.0f,  0.0f,  0.0f}}},
+      {{ 0.0f,  0.5f, 0.0f}, {{ 0.0f,  0.0f,  0.0f}}},
+      {{-1.0f, -0.5f, 0.0f}, {{ 0.5f,  0.0f,  0.0f}}},
+      {{ 1.0f,  0.8f, 0.0f}},
+      {{ 0.8f,  1.0f, 0.0f}},
     };
     // clang-format on
 
@@ -33,10 +36,18 @@ class Triangle : public vs::Application {
     glGenBuffers(1, &vbo);
     glBindBuffer(GL_ARRAY_BUFFER, vbo);
 
-    glBufferData(GL_ARRAY_BUFFER, sizeof(vertices), vertices, GL_STATIC_DRAW);
-    glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, 3 * sizeof(float),
+    glBufferData(GL_ARRAY_BUFFER, vertices_.size() * sizeof(vs::Vertex),
+                 vertices_.data(), GL_STATIC_DRAW);
+
+    // Position
+    glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, sizeof(vs::Vertex),
                           (void*)0);
     glEnableVertexAttribArray(0);
+
+    // Color
+    glVertexAttribPointer(1, 3, GL_FLOAT, GL_FALSE, sizeof(vs::Vertex),
+                          (void*)(sizeof(vs::Vec3)));
+    glEnableVertexAttribArray(1);
 
     glBindBuffer(GL_ARRAY_BUFFER, 0);
     glBindVertexArray(0);
@@ -50,11 +61,12 @@ class Triangle : public vs::Application {
 
     glBindVertexArray(vao_);
     shader_.Use();
-    glDrawArrays(GL_TRIANGLES, 0, 3);
+    glDrawArrays(GL_TRIANGLES, 0, (GLsizei)vertices_.size());
   }
 
   unsigned int vao_;
   vs::Shader shader_{};
+  std::vector<vs::Vertex> vertices_;
 };
 
 VS_REGISTER_APP(Triangle);